+++ src/NoteDialog/NoteDetails.styled.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components/macro';
 
-export const NoteDetails = styled.p`
+export const NoteDetailsContainer = styled.p`
   display: flex;
   flex-direction: column;
   margin-bottom: 8px;