+++ src/domain/interfaces/xy-coords.interface.ts
@@ -1,0 +1,4 @@
+export interface XYCoords {
+  x: number;
+  y: number;
+}